pyproject.toml
[project]
name = "gemini-predict"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.8"
dependencies = [
    "fastapi>=0.115.13",
    "google-genai>=1.20.0",
    "python-dotenv>=1.1.0",
    "slack-sdk>=3.35.0",
]
